added mutex locks, and hypercalls for reporting euid and egid

+++ include/linux/hypercall.h
@@ -31,7 +31,7 @@ static inline void igloo_hypercall(uint32_t num, uint32_t arg1) {
 #endif
 }
 
-static inline unsigned long igloo_hypercall2(unsigned long num, unsigned long arg1, unsigned long arg2) {
+static inline unsigned long igloo_hypercall2(uint32_t num, uint32_t arg1, uint32_t arg2) {
 #if defined(CONFIG_ARM)
     register unsigned long r0 asm("r0") = num;
     register unsigned long r1 asm("r1") = arg1;
